Move Serialisable to abc and state its roundtrip laws there
The serialisation interface belongs next to the categorical ones it is
orthogonal to, so Serialisable moves from utils to abc, where the laws
it now states are within reach of the axiom decorator.

The laws move with it, out of axioms.Testable: transparency, pickling
and serialisation are properties of a term that writes itself down, not
of one that generates itself, and the two axes vary independently — a
type that generates its instances need not write itself down, and one
that writes itself down need not be generated. Testable is left the
generation contract alone. Copying joins them: a deep copy goes through
the same reduction as a pickle without the bytes, which is how the
parameters of a NamedGeneric were lost, and no law pinned it.

Collecting an axiom is no longer the business of Category either:
declared_axioms walks the MRO of any class and Category.axioms delegates
to it, so a carrier that states the roundtrips without being a category
is enrolled like the rest. proptest/test_serialisation.py checks every
roundtrip of every such carrier, cat.Ob and cat.Box to begin with, where
the matrix of categories only reached them through the arrows containing
them. BinaryBoxConstructor declares what serialises it without
implementing the interface, which the box it is mixed into provides.
