+++ CONTRIBUTING.md
@@ -78,9 +78,11 @@ cores, with `-p no:benchmark` unloading the benchmark plugin that is
 incompatible with it; drop both to run serially, e.g. when debugging a
 single cell.
 
-Every cell of the matrix is one axiom of one category, named
-`<module>.<Category>.<law>`, so pytest's own `-k` selects cells for
-shorter, targeted tests.
+Every cell of the matrix is one axiom of one carrier, named
+`<module>.<Carrier>.<law>`, so pytest's own `-k` selects cells for
+shorter, targeted tests. Most carriers are categories; the roundtrip
+laws of `discopy.abc.Serialisable` are also checked on the terms that
+state them without being categories, e.g. the objects of a category.
 
 ```shell
 uv run pytest proptest/ -k unitality -v
